move has_patch_ghost test to init
